feat(iota-tool): simplify by reading the existing genesis and deriving the chain id and bucket
