+++ docs/content/operator/common/snapshots.mdx
@@ -258,18 +258,13 @@ A node restored from a formal snapshot holds only the end-of-epoch checkpoint su
 ```shell
 iota-tool backfill-checkpoint-summaries \
    --path "<PATH-TO-NODE-DB>/live" \
-   --genesis "<PATH-TO-GENESIS-BLOB>" \
-   --network mainnet \
    --num-parallel-downloads 50 \
    --verbose
 ```
 
-   - `--path`: The node's live database directory (the one containing `checkpoints/`, `store/`, and `epochs/`).
-   - `--genesis`: The path to the network's `genesis.blob`.
-   - `--network`: Network whose default archive bucket to read from. Defaults to "mainnet".
-   - `--skip-verify`: Skip the pairwise chain verification of the downloaded summaries (verification is on by default).
+   - `--path`: The node's live database directory (the one containing `checkpoints/`, `store/`, and `epochs/`). The network — and thus the archive bucket — is derived from the node's own genesis checkpoint, so no genesis file or network flag is needed. Override the bucket with the `CUSTOM_ARCHIVE_BUCKET` environment variables for a custom network.
 
-This only adds historical summaries below the node's existing watermarks — it never changes how far the node has synced or executed. It is safe to re-run.
+Each downloaded summary is verified against the committee chain before it is stored. This only adds historical summaries below the node's existing watermarks — it never changes how far the node has synced or executed, and it never overwrites the summaries the node already has. It is safe to re-run.
 
 ## IOTA Foundation managed snapshots
